requirements.txt
Django==4.2.20
django-cities-light==3.10.1
django-widget-tweaks==1.5.0
Pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2019.3
sqlparse==0.5.3
weasyprint==60.2
python-barcode==0.15.1 